+++ src/User/UpdateUserCommand.php
@@ -45,6 +45,9 @@ final class UpdateUserCommand
     /** @var bool|null */
     private $isProfessional;
 
+    /** @var string|null */
+    private $intraEuropeanCommunityVAT;
+
     /** @return string|null */
     public function getPhone(): ?string
     {
@@ -223,6 +226,18 @@ final class UpdateUserCommand
     public function setIsProfessional(?bool $isProfessional): self
     {
         $this->isProfessional = $isProfessional;
+
+        return $this;
+    }
+
+    public function getIntraEuropeanCommunityVAT(): ?string
+    {
+        return $this->intraEuropeanCommunityVAT;
+    }
+
+    public function setIntraEuropeanCommunityVAT(?string $intraEuropeanCommunityVAT): self
+    {
+        $this->intraEuropeanCommunityVAT = $intraEuropeanCommunityVAT;
 
         return $this;
     }